requirements.txt
pytest==7.4.3
allure-pytest==2.13.2
selene==2.0.0rc4
requests==2.31.0
selenium~=4.15.2